Implement signature sharing through the coordinator
This implements the connection to the sync server from revaultd, and use
it to share and fetch the revocation transactions signatures in the
'revocationtxs' RPC call.

Emergency transactions are not encrypted until we settle https://github.com/re-vault/practical-revault/issues/72 .

We share and poll transactions asynchronously, obviously it's flawed if
stakeholders are not perfectly in sync to presign the txs. It will be
fixed in the next commit as this one is already becoming quite big.