db: block chain re-organization handling
We have a bulky approach to reorgs.
Since we don't maintain a header chain on our side we can't try to be
smart. We just store the last tip we heard about from bitcoind, so if
for some reason it differs and is not in the future (or it is but our
current one is not valid anymore according to bitcoind) we take all the
vaults in DB and check them back against the best block chain (ie
bitcoind).
If a deposit which was previously confirmed gets unconfirmed, we drop
all the presigned transactions from the DB.

Unfortunately the post-reorg rescan handling is inherently common to
the bitcoind and database modules, so here we are opening part of our
nicely contained logic to expose transactions taking a database
transaction for processing the entire rescan in a single transaction.

Note how concurrency doesn't help us here too: what if the users decide
to call 'revocationtxs' at the precise time the bitcoind thread detects
a (~malicious) reorg and drops the presigned transactions from the DB.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>