db: add a 'spend_transactions' table
We are going to store Spend transactions so that they can be updated
across restarts (eg if a manager signs the next day we created the Spend
transaction).

This implements the database side of it, including the interface and the
two tables (one for the actual transactions, the other one to bridge
with the Unvault transactions).

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>