db: enable the 'unlock_notify' feature for write accesses
This is the smallest possible patch for fixing our concurrent access to
the SQLite database, albeit not being the most efficient one.

In a nutshell, 'unlock_notify' will basically wait for the database to
be unlocked to start writing to it.

The transaction behaviour was changed too as a deadlock could occur when
upgrading a read lock to a write lock, for example when starting with
read accesses before writing in a single transaction.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>