bitcoind: generalize the 'listunspent' poller interface
It was specific to the deposit descriptor and we are actually going to
use it for the Unvault one as well.

It also did not make sense at all to have the vault status as part of
the UTXO information data structure, we just want to know if it's
confirmed.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>