bitcoind: track confirmation of Spend transactions
In a rather convoluted manner, we retrieve the Spend transaction id from
bitcoind and store it as a new column in the 'vaults' database. This
allows us to then check for each vault in 'spending' state if the
corresponding Spend transaction is confirmed after each tip state.

This is bulky, but being smart is probably not worth it there (yet?)

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>