jsonrpc: fix shutdown handling with mio
Our handling of shutdown on Unix was misguided. We now special-case it as synchronous (and move
the check after the actual handler of a read event, as a single event can be both Read and Read
Closed !) so we are sure to never miss it.

This also waits for all commands handler to end (by joining their
threads) before exiting.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>