revaultd: correct reorg handling
We were previously getting rid of everything and simply marking the
vault as 'unconfirmed' no matter its state when we encountered a deep
enough reorg that it would unconfirm the deposit.

It was intuitive, simple... And wrong. We would not be able to track the
state of the vault anymore without a significant overhaul of the
bitcoind polling logic (basically switching to minconf=1 for listunspent) since
bitcoind's wallet will consider the deposit as spent if there exist a
transaction in the wallet spending them... Even if it's invalid !! (eg
if it's timelocked multiple blocks in the future and can't even be
included in the mempool).

What we do now is marking the vault as 'unconfirmed' only if it never got to
the second layer of the transaction chain (the Unvault was never
broadcast), and marking it as 'unvaulting' if it did get to the second
layer. This results in it passing immediately to 'canceling' or
'spending', which is annoying for 'spending' this it may stay in this
state for awhile because timelock.... But we assume a several-dozens
block reorg isn't likely on mainnet anyways, and if it happens we have
bigger problems than fine-grained status tracking for our vaults.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>