bitcoind: fixups for edgy state transitions between Cancel and Spending
What happens if a vault was 'spending' and the Spend transaction gets
evicted from the mempool? What if it then gets canceled in addition?
Same for the other way around.

We now check at each block (we may want to do it at each poll eventually)
if the vaults that are 'spending' and 'canceling' still have their
transaction in the mempool if they are not yet confirmed. Otherwise we
downgrade them to 'unvaulted'.

Also, we now only acknowledge vault that are 'spending' or 'canceling'
if the transactions are really into the mempool.

Co-authored-by: Daniela Brozzoni <daniela@revault.dev>
Co-authored-by: Antoine Poinsot <darosior@protonmail.com>