tests: integrate stakeholder-managers to the testing framework
We were formerly only testing against well separated roles, and this
bit us.

I previously thought we could just have some of the StakeholderRevaultD
also be manager but it'd be actually pretty convoluted and not in phase
with our use in the tests. Therefore, i just went with complete
duplication: there is now a third role to stakehodlers and managers:
stakeholders-managers.