bitcoind: always recover vault status
We would previously not retrieve the right state of a deposit if we did
not actively watch it. That was purely dumb: what if it moved while we
were away?

I also noticed we only checked for deposits recovery while we were not
actively watching, but not for vault state recovery! This therefore adds
a comprehensive sanity check that we could retrieve any of the state.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>