tests: remove unnecessary waits for transient logs
We were waiting for each transient state before waiting for the final
one in the routines. This is unnecessary, if revaultd isn't robust
enough that we need to take care of waiting before generating blocks
there is an issue that must be solved.

Of course, we still test each state transition in the appropriate test
but no need to repeat ourselves everywhere.

This also does cleanup for DRY after all the rushed fixups for the demo
deadline.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>