bitcoind: unload any leftover watchonly wallet at startup
We used to have this check and i removed it since it does not seem right
for us to unload wallets on the host's bitcoind. But it's actually fine
if it's the very same one we create (by name) and we'll now also refuse
to start with an inconsistent number of loaded watchonly wallets.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>