db: startup sanity check of descriptors against config
We need the descriptors to be present in database (obviously) for state
consistency, and we need them to be in the config for first startup. We
are going with this duplication anyways so the best we can do is to
error early if they diverge as it wouldn't smell good.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>