mscompiler: do some sanity check on descriptor before returning
This is unlikely, but seems am so lazy that i actually copy pasted
static pubkeys derived from the stakeholders xpubs to use them for
cosigning servers in the configuration example Edouard experienced a
crash when deriving the first set of pre-signed transactions!

So, better safe than sorry, check the descriptors before letting the user
shoot themselves in the foot.

Reported-by: Edouard Paris <edouard@revault.dev>
Signed-off-by: Antoine Poinsot <darosior@protonmail.com>