qa: wait_for, don't immediately assert in test_reorged_deposit
We were asserting right after reading the log line. Thing is, the log
line is emitted inside a database tx. So we could query `listvaults` and
get the response prior to the db tx to be written. This caused a race
making the test flaky.

Also, this blacks out the tests... I really need to add a linter check
to the CI.

Signed-off-by: Antoine Poinsot <darosior@protonmail.com>