Panic instead of process::exit
In order to exit the program due to an unrecoverable error,
the recommend way is to panic!(). This will unwind the stack
(run all destructors) and exit the program with additional
information (including the message string you can specify).

see:
https://stackoverflow.com/questions/39228685/when-is-stdprocessexit-o-k-to-use

Revault-gui starts Revaultd into a child thread.
We want the child thread to panic but not the main thread to exit.