Import CPFP descriptor into Core and... ...cpfp_secret in revaultd
The CPFP private key can be stored in a file called
`cpfp_secret` in the datadir. If the `cpfp_secret` file
is not found, CPFP will be disabled. If it's found with an
invalid key or a key not present in the descriptor, we error.
The CPFP descriptor is imported directly in Bitcoin Core,
as the revault_tx update modifies it to be a `multi`. This
commit introduces code for creating a new wallet, and some
functions for querying it.
We also create wallets with blank=true, as we don't want them to
have private keys.
This commit also renames the `KeyError` enum to `NoiseKeyError`,
and moves the "Not a fresh wallet..." debug message out of the loop,
so we don't spam the logs.