CPFP transactions when needed
At every new tip, checks if we have unvaults/spends that have
priority and are still unconfirmed. If yes, broadcasts a CPFP
transaction.
There's no need to explicitly check if we already CPFPed a certain
transaction, as if we did, our new CPFP will either RBF the previous
one or won't be accepted by the mempool