Presigned txs signatures update overhaul
This is a refactoring of the awful and convoluted logic of
db_update_presigned_tx as used in the signature fetcher in
view of the watchtower integration.

This separates the computation of the next vault status and the
transaction update, refactors the merge of new signatures with
in-DB transactions and batches the update of per-vault transactions.

This then refactors the DB transaction getters that were tailored
to the db_update_presigned_tx need (returning a DB is and the inner
PSBT). It instead now returns the DbTransaction, which was added a
set of methods to deal with the inner PSBT in place of the hacky
assert_tx_type macro that was handed around.
I tried hard but in the end i faced the fact that i won't find a
generalistic solution for the `RevaultTransaction`s in the DbTransaction
and i have to live with the duplication of matches.

As a consequence, this refactored the update of signatures to use
revault_tx's add_signature(). This meant that we could get rid of all
our duplicated signature-checking logic (Fixes #286).

As i had to refactor a lot of the RPC commands as a consequence of the
refactoring, this also starts tackling #145 by returning JSONRPC
internal errors instead of crashes in super-edgy cases where we polled
the status of vault, did some stuff outside of the DB transaction, and
assumed the status of the vault to still be the same and query the
corresponding presigned transactions.