qa: introduce a proxy for bitcoind to allow to mock some RPC calls
This makes the revaultds and miradords hit a server we set up instead of
bitcoind. This server will either pass the calls through to bitcoind or
respond with some custom mocks.

You must opt-in to it at deployment time.
There seems to be a significant performance hit if you do.

This contains code stolen from C-lightning: the proxy code using Flask +
cheroot.