bitcoind: base the decision to CPFP on each tx individual feerate
This was incorrectly considering the set of unvaults (or spends) as "a
package". It's misleading as a package is usually used to refer to
linked transactions in Bitcoin (mempool) land.
Maybe this confusion caused the error: it considered the entire set of
transactions' feerate as if they somehow had an influence on each other
wrt mining.

Instead, filter transactions to be bumped individually. In addition we
introduce a threshold below which we don't feebump (probably not worth
paying for a CPFP if you are like 2sat/vb below the next block
feerate!).