jsonrpc: move all commands' implementation to a new 'command' module
This is the first step to a Rust library API in addition to the JSONRPC
one.
This unfortunately makes for a large and not moveonly commit. This tries
to reasonably conserve behaviour, but still contains some adaptations
and tiny drive-by fixes.
This introduces some ugly circular dependencies between command.rs and
control.rs, which should be fixed by merging them into a single module.