lib: split daemon_main in daemon_start and rpc_server_loop
This allows a client of the library to start the daemon without having
to start the RPC server.
This creates a thread similar to bitcoind for the sigfetcher for
aesthetic purposes (although it could turn useful in the future if we
extend the messages of the sigfetcher or if we want to mock it).