Rework the rescan
Let's say we have a reorg at h = 106, creating a chain split: the chain
107a, 108a... and 107b, 108b...
Our latest db tip was 108a. We unwind the state of the db to 106, set
the db tip to 107, let update_utxos do its job with listsinceblock from
the block 107.
This could work, but in some cases we "skip" states going, for example,
from unvaulting to canceling, not setting the stage blockheight.
Also this needs more unit tests.