Clean jsonrpc error module
Since jsonrpc server uses the daemon commands
interface, the errors are handled by the
unique enum CommandError and all the helpers
in the jsonrpc api error module are deprecated.

The ErrorCode enum is moved to the command module
in order to be exposed for the consumers importing
the lib without the jsonrcp server feature.