poller: track confirmation of Unvault tx even for not 'unvaulting' vaults
We can currently pass from 'unvaulting' to 'spending' state. This was
previously not an issue, but we now require the first stage tx (here the
unvault) block height to be set for correct reorg handling.

Therefore, for a vault whose Unvault tx is spent without being confirmed
first check whether it's confirmed when it's spent and then at every
block until it's spent.