poller: always check if a transaction is still active
When checking if a spending transaction would have confirmed, we would
bypass the 'getmempoolentry' on an error to call 'gettransaction'. We
shouldn't, as if for some reason our transaction is dropped from the
watchonly wallet and another version confirms (for instance another
Spend, or another Cancel once we have multiple of them) then we'd stay
stuck here.

This doesn't matter for now, but might prevent some sneaky bugs in the
future.