Refac bitcoind request retry after startup
Code can not prevent Connection and I/O Errors to happen, in that case
the bitcoind client can retry multiple time its request in the given limit.
This commit implements the try limit and make the request try again
in case of i/o or timeout errors.

In order to fail faster at startup the try limit is set after the
bitcoind sanity checks.

close #372