Add deposit_amount to ListSpendTxEntry
Psbt of the spend transaction is consuming
unvault outputs and do not contains information
about the total amount of the vaults moved.
Then Api client needs to retrieve for each spend transaction
the vaults to calculate the total.
Because the command is already retrieving the vaults
to get the outpoints, the calcul can be done at this level.