commands: remove the unit test of listpresigned txs
It's not that useful as listpresignedtxs is already exercised under all
conditions by the functional tests. And it would be a MAJOR pain to move
it to the multi-cancel logic. As the usefulness / time spent ratio is
very low, just drop it.

I'd still like to add it back if possible, but definitely not with
hardcoded PSBTs.