bitcoind: check for the Unvault's spender from the Unvault conf height
This bug was found by reducing the bitcoind poll interval in the
functional tests.

We were re-using the previous tip hash as the last block to poll an
Unvault's spender from.
But a deposit may be noticed as spent but not marked as such if it was not
yet marked as active by the signature fetcher thread (it is in order to
keep a consistent state as the sigfetcher might otherwise overwrite it).
In this case we would update the `previous_tip` to the next tip (and never
check for this block again) and never notice the Unvault transaction spender
if it was there.