qa: reliable censoring reorgs
We would previously deprioritise transactions from the mempool before
generating to mock the censoring behaviour. But it would fall short if
the transactions got evicted during the reorg and re-submitted by
revaultd!

Instead, opt for a reliable way of making sure we mine blocks without a
single transaction in it.